feature : paypoint와 member의 point Integer로 수정2

+++ backend/member/src/main/java/com/paymong/member/paypoint/dto/AddPaypointResDto.java
@@ -10,6 +10,6 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class AddPaypointResDto {
-    Long point;
+    Integer point;
     String mapCode;
 }
